% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbay_aggregate.R
\name{bitbay_aggregate}
\alias{bitbay_aggregate}
\title{Aggregate BitBay transactions}
\usage{
bitbay_aggregate(data, aggr_time = 1 * 60)
}
\arguments{
\item{data}{object of class \code{bitbay_transactions} generated by
\code{bitbay_trades} or \code{bitbay_trades_from_date} functions.}

\item{aggr_time}{a numeric vector of length 1 with seconds}
}
\value{
A time series object of class \code{zoo} and \code{xts} with
following columns:
\code{Open}, \code{High}, \code{Low}, \code{Close}, \code{Mean}, \code{Median},
\code{Volume_total}, \code{volume_buy}, \code{Volume_sell}, \code{Count}

(Each column name has a \code{coin_currency.} prefix)
}
\description{
\code{bitbay_aggregate} aggregates BitBay transactions by
\code{aggr_time}-time interval. If in a given time interval there are no
transactions (count is 0) then the last available "Close" value is used.
}
\examples{

\dontrun{
data <- bitbay_trades(coin = "BTC", currency = "USD", last_trades = 200)

aggr_data_1min <- bitbay_aggregate(data, aggr_time = 60)
aggr_data_15min <- bitbay_aggregate(data, aggr_time = 15 * 60)
aggr_data_1h <- bitbay_aggregate(data, aggr_time = 60 * 60)
}
}
\references{
BitBay:
\url{https://bitbay.net/en/home}

BitBay Public API:
\url{https://bitbay.net/en/api-publiczne}
}
\seealso{
\code{\link{bitbay_orderbook}}, \code{\link{bitbay_trades}},
         \code{\link{bitbay_trades_from_date}}\code{\link{bitbay_profit}}
}
\author{
Michal Majka
}
