% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbay_profit.R
\name{bitbay_profit}
\alias{bitbay_profit}
\title{Perform a theoretical/virtual market order on the BitBay crypto-currency exhange}
\usage{
bitbay_profit(coin = "BTC", amount = 1, currency = "USD", fee = 0.003,
  investment = NULL)
}
\arguments{
\item{coin}{a character string with a name of the coin.
Available coins: "BTC", "ETC", "LSK", "LTC", "GAME" "DASH" "BCC"}

\item{amount}{a numeric vector of length 1 with the amount of the
crypto currency to be sold}

\item{currency}{a character string with a name of the currency.
Available currencies: "USD", "EUR", "PLN"}

\item{fee}{a numeric vector of length 1 with transaction fee. By default, fee = 0.003 (0.3\%)}

\item{investment}{(optional) a numeric vector of length 1 with the amount of investment}
}
\value{
A list with two data frames - \code{transactions} and \code{summary}.
The former contains transactions and the latter summary (amount sold, profit, etc)
}
\description{
\code{bitbay_profit} function performs a theoretical/virtual market order, where
 a specified amount of crypto-currency is sold at the current market price.
}
\examples{
# Sell 1 BTC for USD with fee 0.3\% and compute profit (investment = 5000 USD)
bitbay_profit(coin = "BTC", amount = 1, currency = "USD",
              investment = 5000, fee = 0.003)

}
\references{
BitBay:
\url{https://bitbay.net/en/home}

BitBay Public API:
\url{https://bitbay.net/en/api-publiczne}
}
\seealso{
\code{\link{bitbay_orderbook}}, \code{\link{bitbay_trades}},
         \code{\link{bitbay_trades_from_date}}
}
\author{
Michal Majka
}
