% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbay_ticker.R
\name{bitbay_ticker}
\alias{bitbay_ticker}
\title{Download ticker from the BitBay crypto-currency exhange}
\usage{
bitbay_ticker(coin = "BTC", currency = "USD")
}
\arguments{
\item{coin}{a character string with a name of the coin.
Available coins: "BTC", "ETC", "LSK", "LTC", "GAME" "DASH" "BCC"}

\item{currency}{a character string with a name of the currency.
Available currencies: "USD", "EUR", "PLN"}
}
\value{
A data frame with two columns \code{Statistic} and \code{Value},
in which rows correspond to the following statistics:

  \code{max} is the rate of transaction, which had highest value

  \code{min} is the rate of transaction, which had lowest value

  \code{bid} is the most profitable rate of active purchase orders

  \code{ask} is the most profitable rate of active sell orders

  \code{vwap} is the weighted average from last 24 hours

  \code{average} is the average rate of 3 best sell orders

  \code{volume} is the number of shares traded during last 24 hours
}
\description{
This function can be used to download the ticker from BitBay market
through the public API.
}
\examples{
ticker_BTC_USD <- bitbay_ticker(coin = "BTC", currency = "USD")
ticker_BTC_USD
attr(ticker_BTC_USD, "pair")
attr(ticker_BTC_USD, "download_time")


ticker_BTC_EUR <- bitbay_ticker(coin = "BTC", currency = "EUR")
ticker_BTC_EUR
attributes(ticker_BTC_EUR)[c("pair", "download_time")]


ticker_LSK_PLN <- bitbay_ticker(coin = "LSK", currency = "PLN")
ticker_LSK_PLN
attributes(ticker_LSK_PLN)[4:5]

}
\references{
BitBay:
\url{https://bitbay.net/en/home}

BitBay Public API:
\url{https://bitbay.net/en/api-publiczne}
}
\seealso{
\code{\link{bitbay_orderbook}}, \code{\link{bitbay_trades}},
         \code{\link{bitbay_trades_from_date}}, \code{\link{bitbay_profit}},
}
\author{
Michal Majka
}
