% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbay_trades.R
\name{bitbay_trades}
\alias{bitbay_trades}
\title{Download transactions from the BitBay crypto-currency exhange}
\usage{
bitbay_trades(pair = "BTC/EUR", last_trades = 50)
}
\arguments{
\item{pair}{a character string with a pair.
Available pairs: combinations of "BTC", "ETC", "LSK", "LTC", "GAME" "DASH" "BCC" and "PLN", "EUR", "USD", "BTC". They have to be separated by "/".}

\item{last_trades}{a positive integer N or a character string "all". If an integer is specified then the last N transactions are going to be downloaded. If "all" then all historic transactions are going to be downloaded.}
}
\value{
A data frame with following four columns:

  \code{Date} date of transaction in POSIXct format (GMT time zone)

  \code{Price} rate

  \code{Order} type of the order (sell/buy)

  \code{Volume} amount of cryptocurrency in that transaction

  \code{TID} unique transaction id
}
\description{
This function can be used to download last N or all historic transactions
from BitBay market through the public API.
}
\examples{

\dontrun{

# Download last 50 trades
last_50_trades <- bitbay_trades()
head(last_50_trades)
tail(last_50_trades)


# Additional attributes (pair and download time)
attributes(last_50_trades)$pair
attributes(last_50_trades)$download_time


trades_LSK_EUR <- bitbay_trades(pair = "LSK/EUR", last_trades = 100)
head(trades_LSK_EUR)
tail(trades_LSK_EUR)
attributes(trades_LSK_EUR)[c("pair", "download_time")]



all_trades <- bitbay_trades(pair = "BTC/USD", last_trades = "all")
head(all_trades)
tail(all_trades)
}

}
\references{
BitBay:
\url{https://bitbay.net/en/home}

BitBay Public API:
\url{https://bitbay.net/en/api-publiczne}
}
\seealso{
\code{\link{bitbay_orderbook}}, \code{\link{bitbay_ticker}},
         \code{\link{bitbay_trades_from_date}}, \code{\link{bitbay_profit}}
}
\author{
Michal Majka
}
