% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbay_orderbook.R
\name{bitbay_orderbook}
\alias{bitbay_orderbook}
\title{Download orderbook from the BitBay crypto-currency exhange}
\usage{
bitbay_orderbook(coin = "BTC", currency = "USD")
}
\arguments{
\item{coin}{a character string with a name of the coin.
Available coins: "BTC", "ETC", "LSK", "LTC", "GAME" "DASH" "BCC"}

\item{currency}{a character string with a name of the currency.
Available currencies: "USD", "EUR", "PLN"}
}
\value{
A list with two data frames containing \code{bids} and \code{asks},
}
\description{
This function can be used to download the orderbook from BitBay market
through the public API.
}
\examples{
orderbook_BTC_USD <- bitbay_orderbook(coin = "BTC", currency = "USD")
head(orderbook_BTC_USD$bids)
head(orderbook_BTC_USD$asks)

}
\references{
BitBay:
\url{https://bitbay.net/en/home}

BitBay Public API:
\url{https://bitbay.net/en/api-publiczne}
}
\seealso{
\code{\link{bitbay_orderbook}}, \code{\link{bitbay_trades}},
         \code{\link{bitbay_trades_from_date}}, \code{\link{bitbay_profit}},
}
\author{
Michal Majka
}
